Added styles for div.container and body
To prevent scrolling, I set the body's width and
height to the full length and height of the viewport.

I also set the display of div.container to flex, and
centered it horizontally within the page. Also
added a border around div.container so the user
knows where the boundaries of the etch-a-sketch are.