Added several new functions
removeDivs(): Removes all divs of the .row class

addDivs(): this function was originally called
createDivGrid, but I changed the name for
consistency. I added a parameter to this function,
'dimensions', so that it can dynamically
create and add any number of divs.

changeGrid(): Uses 'prompt' to get an input from
the user. I implemented input validation so the
user can't enter a string or a number greater than
100 or less than 1. Upon successful validation,
it will call removeDivs and addDivs, and supply
the user's input as an argument to addDivs.